Exp: trying out nonclusted index on answer
trying to work on the nonclusted keys

The down side of using nonclusted keys is that there will be more storage used
and that the inport time will be longer by about 30 sec.
The up side is that the query time will not be that effected by the increase of data in answers.

testing with 2 sets of sav files reveals that the query time will increase by 1 sec with out index.